fix: address chatgpt-codex-connector[bot] follow-up (3 P2 findings)
- **`use-v2-exchange-config` + `use-rebalance-check` — bounded retry
  doesn't recover from longer outages**: SWR 2.x suppresses
  `refreshInterval` revalidation while the hook is in error state, so
  a fixed `errorRetryCount` budget (or `shouldRetryOnError: false`)
  leaves the cache wedged after the budget exhausts. Replaced both
  hooks' retry config with a custom `onErrorRetry` that schedules
  the next attempt at the same cadence as `refreshInterval` (60s for
  v2-exchange-config, 30s for rebalance-check) and loops indefinitely
  while the document is visible. Visibility gating prevents background
  tabs from piling up upstream calls during a long outage.

- **`v2-exchange-config` route — RPC URL leak**: viem embeds the
  transport URL in error messages + stacks. Forno doesn't carry
  secrets today, but a future RPC provider with path-based API keys
  (Infura/Alchemy-style) would leak credentials into Sentry/server
  logs through the route's `Sentry.captureException(err, …)` and
  `console.error(err)` paths. Lifted the existing `redactRpcUrl` /
  `containsRpcUrl` helpers out of `rebalance-check/route.ts` into
  `lib/redact-rpc-url.ts` (the rebalance-check route re-exports for
  test back-compat) and now redacts before capture/log. Same defense
  the rebalance-check route already had.
